Move Penguin to PenguinTables.
This clears the way for Penguin to be a package that re-exports all
the other packages.
