+++ Package.swift
@@ -8,8 +8,8 @@ let package = Package(
     products: [
         // Products define the executables and libraries produced by a package, and make them visible to other packages.
         .library(
-            name: "Penguin",
-            targets: ["Penguin"]),
+            name: "PenguinTables",
+            targets: ["PenguinTables"]),
         .library(
             name: "PenguinCSV",
             targets: ["PenguinCSV"]),
@@ -28,19 +28,17 @@ let package = Package(
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
-        // .package(url: /* package url */, from: "1.0.0"),
-        // .package(url: "https://github.com/tensorflow/swift-apis.git", .branch("master")),
         .package(url: "https://github.com/google/swift-benchmark.git", .branch("master")),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages which this package depends on.
         .target(
-            name: "Penguin",
+            name: "PenguinTables",
             dependencies: ["PenguinCSV", "PenguinParallel"]),
         .testTarget(
-            name: "PenguinTests",
-            dependencies: ["Penguin"]),
+            name: "PenguinTablesTests",
+            dependencies: ["PenguinTables"]),
         .target(
             name: "PenguinCSV",
             dependencies: []),
