+++ Sources/PenguinTables/Bool+Comparible.swift
