+++ Sources/PenguinTables/Documentation/Guides/Custom Aggregation.md
