+++ Sources/PenguinTables/Documentation/Sections/Aggregation Internals.md
