+++ Sources/PenguinTables/DoubleConvertible.swift
