+++ Sources/PenguinTables/Error.swift
