+++ Sources/PenguinTables/StringParsible.swift
