+++ Sources/PenguinTables/ThreeWayComparison.swift
