+++ Sources/PenguinTables/TypedColumn.swift
