+++ Tests/PenguinTablesTests/ColumnTests.swift
@@ -14,7 +14,7 @@
 
 import XCTest
 
-@testable import Penguin
+@testable import PenguinTables
 
 final class ColumnTests: XCTestCase {
   func testIntConversion() {
