+++ Tests/PenguinTablesTests/PenguinTests.swift
@@ -12,33 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import PenguinCSV
 import XCTest
 
-@testable import Penguin
+@testable import PenguinTables
 
-final class CSVParsibleTests: XCTestCase {
-  func testIntParsing() throws {
-    assertParse(" 1", as: 1)
-    assertParse("0", as: 0)
-    assertParse(" 100 ", as: 100)
-    assertParse(" -123", as: -123)
+final class PenguinTests: XCTestCase {
+  func testExample() {
+    // This is an example of a functional test case.
+    // Use XCTAssert and related functions to verify your tests produce the correct
+    // results.
+    XCTAssertEqual(Penguin().text, "Hello, World!")
   }
 
   static var allTests = [
-    ("testIntParsing", testIntParsing)
+    ("testExample", testExample)
   ]
-}
-
-fileprivate func assertParse<T: PCSVParsible & Equatable>(
-  _ bytes: String,
-  as val: T,
-  file: StaticString = #file,
-  line: UInt = #line
-) {
-  var s = bytes
-  s.withUTF8 { s in
-    let parsed = T(CSVCell.raw(s))
-    XCTAssertEqual(parsed, val, file: file, line: line)
-  }
 }
